remove `__unused` suffix
This is a non standardized language extension. C23 introduced a similar attribute, `[[maybe_unused]]` [1]. GCC comes with an extension that specifies `__attribute__((unused))` [2]. Other forms like `__unused__`
are found in the wild, often transformed to the GCC form using macros:

```c
```

The utilization here seems to be non-compliant to the C language, and does not compile on GCC 12.2.0 nor on ccomp. Thus we remove the suffix entirely.

[1] https://en.cppreference.com/w/c/language/attributes/maybe_unused
[2] https://gcc.gnu.org/onlinedocs/gcc-13.2.0/gcc/Common-Variable-Attributes.html